+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Drop support for PHP versions older 7.4 (i.e. PHP 7.1 - 7.3)
 - Drop support for ownCloud versions older than 10.5.0 (i.e. OC 10.0 - 10.4)
 - Drop support for Nextcloud versions older than 18 (i.e. NC 13 - 17)
+- New design including cover art on all list-like views
 - Ampache API:
   * The action `download` doesn't implicitly record the track as played (unlike `stream`)
   * The song property `url` refers to the `stream` URL instead of `download` URL